+++ src/lib/index.ts
@@ -1,3 +1,4 @@
 export { ReactScreenshotTaker } from "./react/ReactScreenshotTaker";
 export { ReactScreenshotTest } from "./react/ReactScreenshotTest";
+export { recordCss } from "./recorded-css";
 export { Viewport } from "./screenshot-renderer/api";
