+++ src/tests/__image_snapshots__/Styled components - Desktop - global CSS blue-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c440ecc4c97b44d6e631471477133dd0fec9d96b8ac7e2f19b095227337f2699
-size 13341
+oid sha256:125b531b04438da7999680235737f8d0c00aa228ceab7f3effe6d6080ed2bb8f
+size 8810
