+++ src/tests/__image_snapshots__/Styled components - iPhone X - global CSS orange-snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c9e2cfb7f0d2c9382b1c2201e89edf14c8cddc1504c376a779d8adb33c07a323
-size 52950
+oid sha256:648790129ceff327e4ae182f61f475b325452c21088f3da26736df8106b401a3
+size 34557
